move benchmarks to separate dir
