+++ src/benchmarks/contracts/func/jetton/op-codes.fc
