+++ src/benchmarks/contracts/func/notcoin/gas.fc
