+++ src/benchmarks/contracts/func/notcoin/stdlib-custom.fc
