+++ src/benchmarks/contracts/func/notcoin/workchain.fc
