+++ src/benchmarks/contracts/jetton_minter_discoverable.tact
