+++ src/benchmarks/escrow/escrow.spec.ts
@@ -21,7 +21,7 @@ import {
 } from "../util";
 import benchmarkResults from "./results_gas.json";
 import { readFileSync } from "fs";
-import { posixNormalize } from "../../../utils/filePath";
+import { posixNormalize } from "../../utils/filePath";
 import { resolve } from "path";
 import benchmarkCodeSizeResults from "./results_code_size.json";
