+++ src/benchmarks/jetton/jetton.spec.ts
@@ -39,8 +39,8 @@ import benchmarkResults from "./results_gas.json";
 import { join, resolve } from "path";
 import { readFileSync } from "fs";
 import { storeProvideWalletAddress } from "../contracts/output/escrow_Escrow";
-import { posixNormalize } from "../../../utils/filePath";
-import { type Step, writeLog } from "../../utils/write-vm-log";
+import { posixNormalize } from "../../utils/filePath";
+import { type Step, writeLog } from "../../test/utils/write-vm-log";
 
 const loadFunCJettonsBoc = () => {
     const bocMinter = readFileSync(
