+++ src/test/gas-consumption/contracts/benchmark_functions.tact
