+++ src/test/gas-consumption/contracts/benchmark_sha256_as_slice.tact
