+++ src/test/gas-consumption/contracts/benchmark_sha256_big.tact
