+++ src/test/gas-consumption/contracts/codeOf.tact
