+++ src/test/gas-consumption/contracts/deploy.tact
