+++ src/test/gas-consumption/contracts/forward.tact
