+++ src/test/gas-consumption/contracts/functions.tact
